chore(ui): move code back to interaction.ts
