+++ apps/ui/src/models/swim/interaction.ts
@@ -1,4 +1,5 @@
 import type { Keypair } from "@solana/web3.js";
+import type Decimal from "decimal.js";
 
 import type { EcosystemId, Env } from "../../config";
 import type { ReadonlyRecord } from "../../utils";
@@ -30,7 +31,7 @@ export const INTERACTION_GROUPS = [
   INTERACTION_GROUP_REMOVE,
 ];
 
-export interface BaseInteractionSpec {
+interface BaseInteractionSpec {
   readonly type: InteractionType;
   /** Should be overriden when extending this type */
   readonly params: unknown;
@@ -91,7 +92,7 @@ export type InteractionSpec =
   | RemoveExactBurnInteractionSpec
   | RemoveExactOutputInteractionSpec;
 
-export interface BaseInteraction {
+interface BaseInteraction {
   readonly id: string;
   // TODO: Make less redundant with poolId on non-Swap InteractionSpecs
   readonly poolIds: readonly string[];
@@ -129,3 +130,36 @@ export type Interaction =
   | RemoveExactBurnInteraction
   | RemoveExactOutputInteraction
   | SwapInteraction;
+
+// V2 for Pool Restructure
+interface TokenTransferDetail {
+  readonly tokenId: string;
+  readonly ecosystemId: EcosystemId;
+  readonly value: Decimal;
+}
+
+export interface SwapInteractionSpecV2 extends BaseInteractionSpec {
+  readonly type: InteractionType.SwapV2;
+  readonly params: {
+    readonly fromTokenDetail: TokenTransferDetail;
+    readonly toTokenDetail: TokenTransferDetail;
+  };
+}
+
+export type InteractionSpecV2 =
+  | SwapInteractionSpecV2
+  | AddInteractionSpec
+  | RemoveUniformInteractionSpec
+  | RemoveExactBurnInteractionSpec
+  | RemoveExactOutputInteractionSpec;
+
+export interface SwapInteractionV2
+  extends BaseInteraction,
+    SwapInteractionSpecV2 {}
+
+export type InteractionV2 =
+  | AddInteraction
+  | RemoveUniformInteraction
+  | RemoveExactBurnInteraction
+  | RemoveExactOutputInteraction
+  | SwapInteractionV2;
